백엔드 코드리뷰 #2

See merge request cloud_track/class_05/web_project2/team02/sdz!44

+++ src/main/java/com/elice/sdz/category/service/CategoryService.java
@@ -1,60 +1,69 @@
 package com.elice.sdz.category.service;
 
 import com.elice.sdz.category.dto.CategoryRequestDTO;
+import com.elice.sdz.category.dto.CategoryResponseDTO;
 import com.elice.sdz.category.entity.Category;
 import com.elice.sdz.category.repository.CategoryRepository;
-import com.elice.sdz.global.exception.category.CategoryAlreadyExistsException;
-import com.elice.sdz.global.exception.category.CategoryNotFoundException;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.elice.sdz.global.exception.CustomException;
+import com.elice.sdz.global.exception.ErrorCode;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class CategoryService {
 
+    private final long CAPACITY = 5;
     private final CategoryRepository categoryRepository;
 
-    @Autowired
-    public CategoryService(CategoryRepository categoryRepository) {
-        this.categoryRepository = categoryRepository;
-    }
+    public CategoryResponseDTO createCategory(CategoryRequestDTO categoryRequestDTO) {
+        long categoryCount = categoryRepository.count();
+        Category newCategory = categoryRequestDTO.toEntity();
 
-    public Category createCategory(Category category) {
-
-        if (categoryRepository.findByCategoryName(category.getCategoryName()).isPresent()) {
-            throw new CategoryAlreadyExistsException("카테고리 이름이 이미 존재합니다 : " + category.getCategoryName());
+        if (categoryRepository.findByCategoryName(newCategory.getCategoryName()).isPresent()) {
+            throw new CustomException(ErrorCode.CATEGORY_ALREADY_EXISTS);
         }
 
-        return categoryRepository.save(category);
+        if (categoryCount >= CAPACITY) {
+            throw new CustomException(ErrorCode.CATEGORY_CAPACITY_EXCEEDED);
+        }
+
+        return categoryRepository.save(newCategory).toResponseDTO();
     }
 
-    public List<Category> getAllCategories() {
-        return categoryRepository.findAll();
+    public List<CategoryResponseDTO> getAllCategories() {
+        return categoryRepository.findAll().stream().map(Category::toResponseDTO).toList();
     }
 
-    public Category getCategoryById(Long categoryId) {
-        return categoryRepository.findById(categoryId)
-                .orElseThrow(() -> new CategoryNotFoundException("카테고리가 존재하지 않습니다 : " + categoryId));
+    public CategoryResponseDTO getCategoryById(Long categoryId) {
+        Category category = categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new CustomException(ErrorCode.CATEGORY_NOT_FOUND));
+
+        return category.toResponseDTO();
     }
 
-    public Category updateCategory(Long categoryId, Category category) {
+    public CategoryResponseDTO updateCategory(Long categoryId, CategoryRequestDTO categoryRequestDTO) {
         Category existingCategory = categoryRepository.findById(categoryId)
-                .orElseThrow(() -> new CategoryNotFoundException("카테고리가 존재하지 않습니다 : " + categoryId));
+                .orElseThrow(() -> new CustomException(ErrorCode.CATEGORY_NOT_FOUND));
 
-        if (categoryRepository.findByCategoryName(category.getCategoryName()).isPresent()) {
-            throw new CategoryAlreadyExistsException("카테고리 이름이 이미 존재합니다 : " + category.getCategoryName());
+        if (categoryRepository.findByCategoryName(categoryRequestDTO.getCategoryName()).isPresent()) {
+            throw new CustomException(ErrorCode.CATEGORY_ALREADY_EXISTS);
         }
 
-        existingCategory.setCategoryName(category.getCategoryName());
+        existingCategory.setCategoryName(categoryRequestDTO.getCategoryName());
 
-        return categoryRepository.save(existingCategory);
+        return categoryRepository.save(existingCategory).toResponseDTO();
     }
 
     public void deleteCategory(Long categoryId) {
-        categoryRepository.findById(categoryId)
-                .orElseThrow(() -> new CategoryNotFoundException("카테고리가 존재하지 않습니다 : " + categoryId));
+        Category category = categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new CustomException(ErrorCode.CATEGORY_NOT_FOUND));
+
+        if (!category.getProducts().isEmpty()) {
+            throw new CustomException(ErrorCode.CATEGORY_WITH_PRODUCTS);
+        }
 
         categoryRepository.deleteById(categoryId);
     }
